requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dateutil
streamlit-authenticator
PyYAML
bcrypt
